cluster: Only recreate pods if they change
Before, we always deleted and recreated pods when users ran `blimp up`,
which was slow, and caused unnecessary loss of state.

+++ cluster-controller/main.go
@@ -16,6 +16,8 @@ import (
 	rbacv1 "k8s.io/api/rbac/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
@@ -458,13 +460,29 @@ func (s *server) deployCustomerPods(namespace string, desired []corev1.Pod) erro
 
 func (s *server) deployPod(pod corev1.Pod) error {
 	podClient := s.kubeClient.CoreV1().Pods(pod.Namespace)
-	_, err := podClient.Get(pod.Name, metav1.GetOptions{})
+	curr, err := podClient.Get(pod.Name, metav1.GetOptions{})
 	if err != nil && !kerrors.IsNotFound(err) {
 		return fmt.Errorf("get pod: %w", err)
 	}
 
+	applyAnnotation, err := runtime.Encode(unstructured.UnstructuredJSONScheme, &pod)
+	if err != nil {
+		return fmt.Errorf("make apply annotation: %w", err)
+	}
+
+	if pod.Annotations == nil {
+		pod.Annotations = map[string]string{}
+	}
+	pod.Annotations["blimp.appliedObject"] = string(applyAnnotation)
+
 	if err == nil {
-		// TODO: Only do this if the pod changed.
+		// If the currently deployed pod is already up to date, we don't have
+		// to do anything.
+		if curr.Annotations["blimp.appliedObject"] == string(applyAnnotation) {
+			return nil
+		}
+
+		// Delete the existing pod, and before we recreate it.
 		if err := s.deletePod(pod.Namespace, pod.Name); err != nil {
 			return fmt.Errorf("delete pod: %w", err)
 		}
